The Enabled property will now deep-check whether a shortcut also exists within the system when being queried. The DeleteShortcut method will now check for any existing shortcuts before deletion; it will also be called if the UseAlternativeOnFail property is enabled while any target user is set to ensure complete removal of any available shortcuts from the user machine.

+++ BootMeUp/BootMeUp.cs
@@ -157,6 +157,9 @@ namespace WK.Libraries.BootMeUpNS
                         _enabled = true;
                 }
 
+                if (ShortcutExists() && !ShortcutVaries())
+                    _enabled = true;
+
                 return _enabled;
             }
             set {
@@ -497,10 +500,9 @@ namespace WK.Libraries.BootMeUpNS
                 if (UseAlternativeOnFail)
                 {
                     if (TargetUser == TargetUsers.AllUsers)
-                    {
                         Unregister(TargetUsers.CurrentUser);
-                        DeleteShortcut();
-                    }
+
+                    DeleteShortcut();
                 }
             }
             else if (BootArea == BootAreas.StartupFolder)
@@ -746,7 +748,8 @@ namespace WK.Libraries.BootMeUpNS
         {
             try
             {
-                System.IO.File.Delete(ShortcutPath);
+                if (ShortcutExists())
+                    System.IO.File.Delete(ShortcutPath);
 
                 Exception = null;
